# Optimized `Encoder.Encode()`